move ffmpeg health function
